Remove article show container because it is most useful for database extraction and not API calls.
