+++ client/src/actions/types.js
@@ -12,3 +12,5 @@ export const RECEIVE_ARTICLE = 'RECEIVE_ARTICLE';
 export const DELETE_ARTICLE = 'DELETE_ARTICLE';
 export const RECEIVE_ARTICLE_ERRORS = 'RECEIVE_ARTICLE_ERRORS';
 export const RECEIVE_VISIT_PAGE = 'RECEIVE_VISIT_PAGE';
+export const RECEIVE_GET_CREDIBBILITY_SCORE = 'RECEIVE_GET_CREDIBBILITY_SCORE';
+export const RECEIVE_GET_RELIABILITY_SCORE = 'RECEIVE_GET_RELIABILITY_SCORE';
